backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
google-cloud-storage==2.16.0
python-multipart==0.0.9